Added new UI
App Structure:
    - Configured App.tsx to have two main branches, either legacy UI or new UI
    - Added extensions to proccess glsl
    - Removed namespace constraint from eslint
    - Added namespace types for PerlinMaterial type to types.ts
PerlinSphere:
    - Generates a perlin sphere within a canvas
    - Sphere's default color is blue, and although the height and width can be passed as props it BREAKS the shaders
    - Accepts sphere from sphere material
globals.css:
    - Added global transitions
shaders:
    - Shader files are present, but they are not used
    - The material is using string input to the shaders because for some reason importing them onto the material causes errors
Main:
    - Main breaks up the main UI into a grid, where the phone sits on the left side. Then two rows on the right for the other info
Call:
    - This is where the React Three shape lives
    - Mute and hang up buttons also live here
Next Steps:
    - Figure out best way to add transition for phone moving to the center
    - Add mute and hang up buttons
    - Center the Sphere
    - Add color for sphere as prop
    - Add Mobile UI
