+++ gslsl.d.ts
@@ -1,0 +1,4 @@
+declare module '*.glsl' {
+    const content: string;
+    export default content;
+}
