+++ src/app/globals.css
@@ -11,7 +11,26 @@
   :root {
     --background: #0a0a0a;
     --foreground: #ededed;
+    --transition-duration: 200ms;
+    --transition-timing: cubic-bezier(0.4, 0, 0.2, 1);
   }
+}
+
+@layer base {
+  * {
+    @apply transition-all duration-200 ease-in-out;
+  }
+}
+
+.scale-spring {
+  transition: transform 887ms linear(0.00, 0.0198, 0.0876, 0.194, 0.330, 0.486, 0.649, 0.816, 0.974, 1.12, 1.24, 1.34, 1.41, 1.46, 1.47, 1.46, 1.43, 1.38, 1.31, 1.24, 1.17, 1.09, 1.01, 0.947, 0.889, 0.842, 0.808, 0.787, 0.778, 0.781, 0.795, 0.819, 0.850, 0.885, 0.922, 0.959, 0.995, 1.03, 1.06, 1.08, 1.09, 1.10, 1.11, 1.11, 1.10, 1.09, 1.07, 1.06, 1.04, 1.02, 1.00, 0.987, 0.974, 0.963, 0.956, 0.951, 0.949, 0.950, 0.953, 0.958, 0.965, 0.973, 0.982, 0.990, 0.998, 1.01, 1.01, 1.02, 1.02, 1.02, 1.02, 1.02, 1.02, 1.02, 1.02, 1.01, 1.01, 1.00, 1.00, 0.997, 0.994, 0.992, 0.990, 0.989, 0.988, 0.989, 0.989, 0.990, 0.992, 0.994, 0.996, 0.998, 1.00, 1.00, 1.00, 1.00, 1.00, 1.01, 1.01, 1.01, 1.01, 1.00, 1.00, 1.00, 1.00, 1.00, 1.00, 1.00);
+
+}
+
+.scale-snap {
+  transition-property: transform;
+  transition-duration: 80ms;
+  transition-timing-function: cubic-bezier(0, 0.55, 0.45, 1);
 }
 
 body {
