+++ src/app/shaders/fragment.glsl
@@ -1,0 +1,6 @@
+    uniform float uTime;
+    varying float vDisplacement;
+    
+    void main() {
+        gl_FragColor = vec4(vec3(0.1, 0.3, 0.8), 1.0);
+    }
